Move repository for work with external api of film info to repository directory
