+++ app/Providers/MovieInfoProvider.php
@@ -3,8 +3,8 @@
 namespace App\Providers;
 
 use App\Custom\MovieByIdInfoRepository;
-use App\Custom\MovieInfoRepository;
-use App\Custom\OmdbMovieInfoRepository;
+use App\Repositories\FilmApiRepository;
+use App\Repositories\Interfaces\FilmApiRepositoryInterface;
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 use Illuminate\Support\ServiceProvider;
@@ -13,7 +13,7 @@ class MovieInfoProvider extends ServiceProvider
 {
     public function register()
     {
-        $this->app->bind(MovieInfoRepository::class, OmdbMovieInfoRepository::class);
+        $this->app->bind(FilmApiRepositoryInterface::class, FilmApiRepository::class);
         $this->app->bind(ClientInterface::class, Client::class);
     }
